==> 17 Pandas_Essential/materials/notebooks/03_Moving Between Wide and Long Data Formats.ipynb <==
import pandas as pd
import os

scoring = pd.read_pickle(os.path.join("..", "scoring.pickle"))
team_splits = pd.read_pickle(os.path.join("..", "team_splits.pickle"))

scoring.head()

three_years = scoring.loc[(scoring.year > 2000) &
                          (scoring.year < 2004)]

# One record per player/year
three_years = three_years.drop_duplicates(subset=["playerID", "year"])

# Get three random players with complete history between 2001-2003
counts = three_years["playerID"].value_counts()
ids = counts[counts == 3].sample(3).index

# Get actual records corresponding for those players
df3 = three_years.loc[three_years['playerID'].isin(ids)]
basic_df3 = df3[["playerID", "year", "G"]]

basic_df3

pivot = basic_df3.pivot(index="playerID", columns="year", values="G")
pivot

pivot.index

pivot.columns

pivot.index.name = None
pivot.columns.name = None

pivot

pivot.index.name = "playerID"
pivot.columns.name = "year"

basic_df3.set_index("playerID").plot.bar();

pivot.plot.bar();

pivot

pivot = pivot.reset_index()
pivot.columns.name = None
pivot

pivot.melt()


pivot.melt(id_vars="playerID")


pivot.melt(id_vars="playerID", var_name="year", value_name="goals")

pivot.melt(id_vars="playerID", value_vars=[2001, 2002],
           var_name="year", value_name="goals")

larger_df3 = df3[["playerID", "tmID", "year", "G"]]
larger_df3

try:
    larger_df3.pivot(index=["playerID", "tmID"],
                     columns="year", values="G")
except Exception as e:
    print(e)
        

larger_df3 = df3[["playerID", "year", "G", "A"]]
larger_df3

test = larger_df3.pivot(index="playerID", columns="year",
                        values=["G", "A"])
test

test.columns
