==> 17 Pandas_Essential/materials/notebooks/04_Multi-level Indexes.ipynb <==
import pandas as pd
import os

master = pd.read_pickle(os.path.join("..", "master.pickle"))
scoring = pd.read_pickle(os.path.join("..", "scoring.pickle"))
team_splits = pd.read_pickle(os.path.join("..", "team_splits.pickle"))

scoring.head()

mi = scoring.set_index(['playerID', 'year'])
mi.head()

mi.index

len(mi.index.levels)

# same as:
mi.index.nlevels

mi.index.levels[0]

mi.index.levels[1]

mi.groupby(level="year")['G'].max().head()

mi.groupby(level="year")['G'].idxmax().head()

mi.loc[mi.groupby(level="year")['G'].idxmax()].head()

mi.loc[('gretzwa01', 1982), :] 

idx = pd.IndexSlice
mi2 = mi.sort_index()

mi.index.is_lexsorted()

sliced = mi.loc[idx[:, 1997:2000], :]
sliced.head()

sliced = mi.loc[idx['aaltoan01':'adamscr01', 1997:2004:2], :]
sliced.head()

locs = mi.index.get_locs((
    idx['aaltoan01':'adamscr01', 1997:2000]
))
locs

sliced = mi.iloc[locs, :]
sliced.head()

import numpy as np

def get_many_locs(df, slices):
    arr = np.empty(0, dtype="int")
    for s in slices:
        locs = df.index.get_locs((s))
        arr = np.concatenate((arr, locs))
    return arr
    

locs = get_many_locs(
           mi,
           (
               idx['aaltoan01':'adamscr01', 1997:2000],
               idx['aaltoan01':'adamscr01', 2004:2006]
           ))
locs

sliced = mi.iloc[locs, :]
sliced.head(10)

mi = mi.sort_index(level='year')
mi.head()

mi.index.is_lexsorted()

try:   
    mi.loc[idx[:, 1997:2000], :]
except Exception as e:
    print(e)

swapped = mi.swaplevel()
swapped.head()

swapped.index.is_lexsorted()

# Wide and Long Formats Revisited

team_splits.head(5)

team_splits = team_splits.drop(columns=["tmID"])
team_splits = team_splits.set_index(["year", "name"])
team_splits.head()

months = team_splits.columns.map(lambda x: x[:3])
metrics = team_splits.columns.map(lambda x: x[3:])
print(months)
print(metrics)

midx = pd.MultiIndex.from_arrays([months, metrics])
midx

team_splits.columns = midx
team_splits.head()

team_splits = team_splits.stack(level=0)
team_splits.head(10)

team_splits = team_splits.swaplevel(1, 2)
team_splits.head()

team_splits = team_splits.reorder_levels([2,0,1])
team_splits.head()

team_splits.index.levels[2].name = "month"
team_splits.head()

team_splits.groupby(level=["name", "year"]).sum().head(10)

team_splits.unstack(level=["year", "month"]).head()

team_splits.to_pickle(os.path.join("..", "team_splits_mi.pickle"))
mi.to_pickle(os.path.join("..", "scoring_mi.pickle"))
