==> 17 Pandas_Essential/materials/notebooks/06_Getting More Insights into Your Data with Windows.ipynb <==
import pandas as pd
import os

scoring = pd.read_pickle(os.path.join('..', 'scoring_mi.pickle'))
birth_dates = pd.read_pickle(os.path.join('..', 'birth_dates.pickle'))
scoring.head()

wayne = scoring.loc['gretzwa01']
wayne

sums = wayne.groupby(level=0)["G", "A"].sum()
sums

means = sums.rolling(3)["G"].mean()
means.head(10)

%matplotlib inline
sums.assign(roll_mean = means).plot();

means = sums.rolling(3, min_periods=2)["G"].mean()
means.head(10)

means = sums.rolling(3, min_periods=1)["G"].mean()
means.head(10)

means = sums.rolling(3).mean()
means.head()

means = sums.rolling(3, center=True).mean()
means

means = sums.rolling(4, center=True).mean()
means.head()

means = sums.rolling(3, center=True).sum()
means.head()

triang_goals_sum = sums['G'].rolling(3,
                                     win_type='triang',
                                       center=True).sum()
sums.assign(triang = triang_goals_sum).head()

quarter_counts = birth_dates.resample('Q').count()
quarter_counts.rolling(4).sum().head(10)

sums.expanding().sum().plot();

sums.expanding().max().plot();

sums.cummax().plot();

totals = sums.sum()
totals

sums.head()

total_part = sums['G'].expanding().aggregate(lambda x:
                                          x.sum() / totals['G'])
total_part.plot();
